requirements.txt
aiohttp==3.10.9
pydantic==2.9.2
loguru==0.7.2
python-dateutil==2.9.0.post0
setuptools==75.1.0
pytest==8.3.3